refurbd-backend-RAILWAY/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# AI APIs
openai==1.10.0
anthropic==0.18.0

# Payments
stripe==7.10.0

# Email
sendgrid==6.11.0

# Image Processing
Pillow==10.2.0
python-magic==0.4.27

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Redis (caching)
redis==5.0.1

# Location services
geoip2==4.7.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# CORS
fastapi-cors==0.0.6

# File validation
filetype==1.2.0

# PDF generation
reportlab==4.0.9
pypdf==4.0.1

# Logging
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
